LLM_Cluster/streamlit_app/requirements.txt
streamlit==1.28.1
requests==2.31.0
Pillow==10.0.1